refactor(hospital): split R1.4 ghost-path guard into surgery_target_guard.rs
R1.4 (ac42dd8) added `verify_surgery_target_exists` plus its fixture-backed
tests directly into hospital_diagnosis.rs, pushing it from 714/19fn to
847/27fn — past this repository's own god-file threshold (loc > 800 or
functions > 25 with loc > 400). Moves the function, its doc comment, and its
`FixtureRepo`-based tests verbatim into a new sibling module
(`mod surgery_target_guard;`, resolved the same way `doctor_bootstrap` already
is from `doctor_adapter.rs`). Pure code motion, no behavior change, same call
site semantics via `surgery_target_guard::verify_surgery_target_exists`.

Resyncs orchestration/integrations.json's diagnosis.hospital toolchainDigest
to hospital_diagnosis.rs's new sha256 via `code-intel repin --write`, by
literal string replacement so no other pinned content in the file moves.
